TASK#13: Refactoring Takes out thunk functions to separate files
